Faruque

+++ src/pages/admin/CategoryPage.vue
@@ -25,7 +25,7 @@
         <AddCategoryDialog
                 :dialog="dialog  || !!updateCategory "
                 @closeDialog='closeDialog'
-                @emitData='addcategory'
+                @emitData='registerCategory'
                 :updateCategory='updateCategory' 
 
 
@@ -53,6 +53,8 @@ components: {
         ...mapState('category', [
             'categories'
         ]),
+        			...mapState('product', ['products'])
+
     },
 
   data () {
@@ -64,10 +66,16 @@ components: {
   },
    methods: {
      
-          ...mapActions('category', ['addcategory','deleteCategory']),
-          
+          ...mapActions('category', ['addCategory','deleteCategory', 'filterCategoryDatafromDb']),
+          ...mapActions('product', ['filterCategoryDatafromDb']),
+
+          registerCategory(data) {
+            this.addCategory(data)
+          },
           removeCategory(id) {
+            // toLowerCase().includes
             let categoryName = this.categories[id].name
+           this.filterCategoryDatafromDb({value : id})
             this.$q
 					.dialog({
 						title: 'Confirme',
@@ -78,7 +86,22 @@ components: {
 						persistent: true
 					})
 					.onOk(() => {
-							this.deleteCategory(id);
+
+            
+            if(Object.keys(this.products).length == 0) {
+	                  this.deleteCategory(id);
+            }
+            else {
+                this.$q
+					.dialog({
+						title: 'Não Permitido',
+						message: `Não pode apagar a categoria ${categoryName} , por que existem alguns produtos associados a ela.`,
+						cancel: false,
+						ok: false,
+					})
+
+            }
+						
 					});
 
           },
